+++ src/Service/CloudWatchLogs/src/CloudWatchLogsClient.php
@@ -419,6 +419,14 @@ class CloudWatchLogsClient extends AbstractApi
                     'signService' => 'logs',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://logs.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'logs',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://logs.us-isob-east-1.sc2s.sgov.gov',