+++ src/Service/CodeBuild/src/CodeBuildClient.php
@@ -216,6 +216,14 @@ class CodeBuildClient extends AbstractApi
                     'signService' => 'codebuild',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://codebuild.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'codebuild',
+                    'signVersions' => ['v4'],
+                ];
         }
 
         return [