+++ src/Service/DynamoDb/src/DynamoDbClient.php
@@ -1181,6 +1181,14 @@ class DynamoDbClient extends AbstractApi
                     'signService' => 'dynamodb',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://dynamodb.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'dynamodb',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://dynamodb.us-isob-east-1.sc2s.sgov.gov',