+++ src/Service/Firehose/src/FirehoseClient.php
@@ -255,6 +255,14 @@ class FirehoseClient extends AbstractApi
                     'signService' => 'firehose',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://firehose.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'firehose',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://firehose.us-isob-east-1.sc2s.sgov.gov',