+++ src/Service/Kinesis/src/KinesisClient.php
@@ -1625,6 +1625,14 @@ class KinesisClient extends AbstractApi
                     'signService' => 'kinesis',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://kinesis.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'kinesis',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://kinesis.us-isob-east-1.sc2s.sgov.gov',