+++ src/Service/Lambda/src/LambdaClient.php
@@ -653,6 +653,14 @@ class LambdaClient extends AbstractApi
                     'signService' => 'lambda',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://lambda.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'lambda',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://lambda.us-isob-east-1.sc2s.sgov.gov',