+++ src/Service/Scheduler/src/SchedulerClient.php
@@ -377,6 +377,14 @@ class SchedulerClient extends AbstractApi
                     'signService' => 'scheduler',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://scheduler.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'scheduler',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://scheduler.us-isob-east-1.sc2s.sgov.gov',