+++ src/Service/Sns/src/SnsClient.php
@@ -612,6 +612,14 @@ class SnsClient extends AbstractApi
                     'signService' => 'sns',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://sns.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'sns',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://sns.us-isob-east-1.sc2s.sgov.gov',