+++ src/Service/Sqs/src/SqsClient.php
@@ -948,6 +948,14 @@ class SqsClient extends AbstractApi
                     'signService' => 'sqs',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://sqs.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'sqs',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://sqs.us-isob-east-1.sc2s.sgov.gov',