+++ src/Service/StepFunctions/src/StepFunctionsClient.php
@@ -361,6 +361,14 @@ class StepFunctionsClient extends AbstractApi
                     'signService' => 'states',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://states.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'states',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://states.us-isob-east-1.sc2s.sgov.gov',