+++ src/Service/XRay/src/XRayClient.php
@@ -139,6 +139,14 @@ class XRayClient extends AbstractApi
                     'signService' => 'xray',
                     'signVersions' => ['v4'],
                 ];
+            case 'us-isof-east-1':
+            case 'us-isof-south-1':
+                return [
+                    'endpoint' => "https://xray.$region.csp.hci.ic.gov",
+                    'signRegion' => $region,
+                    'signService' => 'xray',
+                    'signVersions' => ['v4'],
+                ];
             case 'us-isob-east-1':
                 return [
                     'endpoint' => 'https://xray.us-isob-east-1.sc2s.sgov.gov',