+++ src/Service/Kms/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: Added `us-isof-east-1` , `us-isof-south-1`, `us-isof-east-1-fips` and `us-isof-south-1-fips` regions
+
 ## 1.7.0
 
 ### Added